refactor: Make xremap_cli()
